Refactor Twitter Container to use the new redux setup
